Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.
This date call works with all variants of date (GNU,BSD,Solaris).

Also do not capture build user+host in this case
to allow to get the same build results anywhere and anytime.